requirements.txt
requests>=2.25.0
tqdm>=4.61.0
numpy>=1.19.0
python-dotenv>=0.19.0
pydantic>=1.9.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
llama-cpp-python>=0.3.0
psutil>=5.8.0
pytest>=6.2.5
httpx>=0.19.0